+++ packages/dataviews/src/bulk-actions-toolbar.tsx
@@ -65,11 +65,12 @@ function ActionTrigger< Item extends AnyItem >( {
 	action,
 	onClick,
 	isBusy,
+	items,
 }: ActionTriggerProps< Item > ) {
 	return (
 		<ToolbarButton
 			disabled={ isBusy }
-			label={ action.label }
+			label={ action.getLabel?.( items ) || action.label }
 			icon={ action.icon }
 			isDestructive={ action.isDestructive }
 			size="compact"
@@ -112,6 +113,7 @@ function ActionButton< Item extends AnyItem >( {
 				setActionInProgress( action.id );
 				action.callback( selectedItems );
 			} }
+			items={ selectedEligibleItems }
 			isBusy={ actionInProgress === action.id }
 		/>
 	);