+++ packages/editor/src/components/post-actions/actions.js
@@ -581,7 +581,15 @@ const viewPostAction = {
 
 const postRevisionsAction = {
 	id: 'view-post-revisions',
-	label: __( 'View revisions' ),
+	getLabel( items ) {
+		const revisionsCount =
+			items[ 0 ]._links?.[ 'version-history' ]?.[ 0 ]?.count ?? 0;
+		return sprintf(
+			/* translators: %s: number of revisions */
+			__( 'View revisions (%s)' ),
+			revisionsCount
+		);
+	},
 	isPrimary: false,
 	isEligible: ( post ) => {
 		if ( post.status === 'trash' ) {