+++ packages/dataviews/CHANGELOG.md
@@ -11,6 +11,9 @@
 ### Internal
 
 -   Remove some unused dependencies ([#62010](https://github.com/WordPress/gutenberg/pull/62010)).
+### Enhancement
+
+-   Add `getLabel` prop in Actions API to support labels that use information from the current item. ([#61941](https://github.com/WordPress/gutenberg/pull/61941)).
 
 ## 1.2.0 (2024-05-16)